+++ backend/canisters/community/impl/src/updates/c2c_join_channel.rs
@@ -14,7 +14,7 @@ use gated_groups::{
     CheckVerifiedCredentialGateArgs,
 };
 use group_chat_core::AddResult;
-use types::{AccessGate, ChannelId, MemberJoined, TimestampMillis, VerifiedCredentialGateArgs};
+use types::{AccessGate, ChannelId, MemberJoined, TimestampMillis, UniquePersonProof, VerifiedCredentialGateArgs};
 
 #[update_msgpack(guard = "caller_is_user_index_or_local_user_index")]
 #[trace]
@@ -32,6 +32,7 @@ async fn c2c_join_channel(args: Args) -> Response {
             is_bot: args.is_bot,
             diamond_membership_expires_at: args.diamond_membership_expires_at,
             verified_credential_args: args.verified_credential_args.clone(),
+            unique_person_proof: args.unique_person_proof.clone(),
         })
         .await
         {
@@ -64,11 +65,24 @@ pub(crate) fn join_channel_synchronously(
     channel_id: ChannelId,
     user_principal: Principal,
     diamond_membership_expires_at: Option<TimestampMillis>,
+    unique_person_proof: Option<UniquePersonProof>,
 ) {
-    match read_state(|state| is_permitted_to_join(channel_id, user_principal, diamond_membership_expires_at, None, state)) {
+    match read_state(|state| {
+        is_permitted_to_join(
+            channel_id,
+            user_principal,
+            diamond_membership_expires_at,
+            unique_person_proof,
+            None,
+            state,
+        )
+    }) {
         Ok(None) => {}
-        Ok(Some(args)) if args.gate.synchronous() => {
-            if !matches!(check_if_passes_gate_synchronously(args), CheckIfPassesGateResult::Success) {
+        Ok(Some((gate, args))) => {
+            if !matches!(
+                check_if_passes_gate_synchronously(gate, args),
+                Some(CheckIfPassesGateResult::Success)
+            ) {
                 return;
             }
         }
@@ -84,11 +98,12 @@ async fn check_gate_then_join_channel(args: &Args) -> Response {
             args.channel_id,
             args.principal,
             args.diamond_membership_expires_at,
+            args.unique_person_proof.clone(),
             args.verified_credential_args.clone(),
             state,
         )
     }) {
-        Ok(Some(check_gate_args)) => match check_if_passes_gate(check_gate_args).await {
+        Ok(Some((gate, check_gate_args))) => match check_if_passes_gate(gate, check_gate_args).await {
             CheckIfPassesGateResult::Success => {}
             CheckIfPassesGateResult::Failed(reason) => return GateCheckFailed(reason),
             CheckIfPassesGateResult::InternalError(error) => return InternalError(error),
@@ -104,9 +119,10 @@ fn is_permitted_to_join(
     channel_id: ChannelId,
     user_principal: Principal,
     diamond_membership_expires_at: Option<TimestampMillis>,
+    unique_person_proof: Option<UniquePersonProof>,
     verified_credential_args: Option<VerifiedCredentialGateArgs>,
     state: &RuntimeState,
-) -> Result<Option<CheckGateArgs>, Response> {
+) -> Result<Option<(AccessGate, CheckGateArgs)>, Response> {
     if state.data.is_frozen() {
         return Err(CommunityFrozen);
     }
@@ -130,19 +146,24 @@ fn is_permitted_to_join(
             } else if channel.chat.members.is_blocked(&member.user_id) {
                 Err(UserBlocked)
             } else {
-                Ok(channel.chat.gate.as_ref().map(|g| CheckGateArgs {
-                    gate: g.clone(),
-                    user_id: member.user_id,
-                    diamond_membership_expires_at,
-                    this_canister: state.env.canister_id(),
-                    verified_credential_args: verified_credential_args.map(|vc| CheckVerifiedCredentialGateArgs {
-                        user_ii_principal: vc.user_ii_principal,
-                        credential_jwt: vc.credential_jwt,
-                        ic_root_key: state.data.ic_root_key.clone(),
-                        ii_canister_id: state.data.internet_identity_canister_id,
-                        ii_origin: vc.ii_origin,
-                    }),
-                    now: state.env.now(),
+                Ok(channel.chat.gate.as_ref().map(|g| {
+                    (
+                        g.clone(),
+                        CheckGateArgs {
+                            user_id: member.user_id,
+                            diamond_membership_expires_at,
+                            this_canister: state.env.canister_id(),
+                            unique_person_proof,
+                            verified_credential_args: verified_credential_args.map(|vc| CheckVerifiedCredentialGateArgs {
+                                user_ii_principal: vc.user_ii_principal,
+                                credential_jwt: vc.credential_jwt,
+                                ic_root_key: state.data.ic_root_key.clone(),
+                                ii_canister_id: state.data.internet_identity_canister_id,
+                                ii_origin: vc.ii_origin,
+                            }),
+                            now: state.env.now(),
+                        },
+                    )
                 }))
             }
         } else {