+++ frontend/app/src/components/home/profile/ChitEvents.svelte
@@ -1,24 +1,20 @@
 <script lang="ts">
     import { OpenChat, type ChitEarned } from "openchat-client";
-    import { getContext, onMount } from "svelte";
+    import { getContext } from "svelte";
     import Calendar from "../../calendar/Calendar.svelte";
     import { isSameDay } from "../../calendar/utils";
     import ChitEventsForDay from "./ChitEventsForDay.svelte";
+    import ChitBalance from "./ChitBalance.svelte";
+    import Toggle from "../../Toggle.svelte";
+    import { i18nKey } from "../../../i18n/i18n";
+    import { chitPopup } from "../../../stores/settings";
 
     const client = getContext<OpenChat>("client");
 
-    $: user = client.user;
-    $: userStore = client.userStore;
-    $: streak = $userStore[$user.userId]?.streak ?? 0;
-    $: balance = $userStore[$user.userId]?.chitBalance ?? 0;
+    $: chitState = client.chitStateStore;
 
     let busy = false;
     let events: ChitEarned[] = [];
-
-    onMount(() => {
-        // no need to do anything with the result explicitly as it will get added to the store automatically
-        client.getUser($user.userId);
-    });
 
     function chitEventsForDay(events: ChitEarned[], date: Date): ChitEarned[] {
         return events.filter((e) => {
@@ -50,23 +46,24 @@
 
 <div class="chit-events">
     <div class="header">
-        {#if streak > 0}
+        {#if $chitState.streak > 0}
             <div class="streak">
                 You are on a
-                <div class="streak-txt">{streak}</div>
+                <div class="streak-txt">{$chitState.streak}</div>
                 day streak!
             </div>
         {/if}
-        {#if balance > 0}
-            <div class="balance">
-                <div class="chit"></div>
-                {`${balance.toLocaleString()} CHIT`}
-            </div>
-        {/if}
+        <ChitBalance size={"large"} me balance={$chitState.chitBalance} />
     </div>
     <Calendar on:dateSelected={(ev) => dateSelected(ev.detail)} {busy} let:day>
         <ChitEventsForDay {day} events={chitEventsForDay(events, day)} />
     </Calendar>
+    <Toggle
+        id={"chit-popup"}
+        small
+        on:change={() => chitPopup.set(!$chitPopup)}
+        label={i18nKey("learnToEarn.showChitPopup")}
+        checked={$chitPopup} />
 </div>
 
 <style lang="scss">
@@ -84,7 +81,7 @@
         border-bottom: none;
         display: flex;
         flex-direction: column;
-        padding-bottom: $sp5;
+        padding: $sp5 0;
         gap: $sp3;
     }
 
@@ -99,25 +96,5 @@
     .streak-txt {
         @include font(bold, normal, fs-160);
         color: var(--accent);
-    }
-
-    .balance {
-        padding: $sp3 $sp4;
-        border-radius: var(--rd);
-        background-color: var(--button-bg);
-        color: var(--button-txt);
-        display: flex;
-        gap: $sp4;
-        align-items: center;
-        width: fit-content;
-        align-self: center;
-        @include font(book, normal, fs-120);
-
-        .chit {
-            background-image: url("/assets/chit.svg");
-            background-repeat: no-repeat;
-            width: $sp5;
-            height: $sp5;
-        }
     }
 </style>