+++ frontend/openchat-agent/src/services/userIndex/candid/idl.js
@@ -43,15 +43,6 @@ export const idlFactory = ({ IDL }) => {
   const ChitLeaderboardResponse = IDL.Variant({
     'Success' : IDL.Vec(ChitUserBalance),
   });
-  const ClaimDailyChitResponse = IDL.Variant({
-    'AlreadyClaimed' : TimestampMillis,
-    'Success' : IDL.Record({
-      'streak' : IDL.Nat16,
-      'chit_earned' : IDL.Nat32,
-      'chit_balance' : IDL.Int32,
-      'next_claim' : TimestampMillis,
-    }),
-  });
   const DiamondMembershipSubscription = IDL.Variant({
     'OneYear' : IDL.Null,
     'ThreeMonths' : IDL.Null,
@@ -90,7 +81,6 @@ export const idlFactory = ({ IDL }) => {
   });
   const CurrentUserResponse = IDL.Variant({
     'Success' : IDL.Record({
-      'streak' : IDL.Nat16,
       'username' : IDL.Text,
       'date_created' : TimestampMillis,
       'is_platform_operator' : IDL.Bool,
@@ -98,12 +88,10 @@ export const idlFactory = ({ IDL }) => {
       'wasm_version' : BuildVersion,
       'icp_account' : AccountIdentifier,
       'referrals' : IDL.Vec(UserId),
-      'next_daily_claim' : TimestampMillis,
       'user_id' : UserId,
       'display_name' : IDL.Opt(IDL.Text),
       'avatar_id' : IDL.Opt(IDL.Nat),
       'moderation_flags_enabled' : IDL.Nat32,
-      'chit_balance' : IDL.Int32,
       'is_suspected_bot' : IDL.Bool,
       'canister_upgrade_status' : CanisterUpgradeStatus,
       'suspension_details' : IDL.Opt(SuspensionDetails),
@@ -359,17 +347,14 @@ export const idlFactory = ({ IDL }) => {
     'volatile' : IDL.Opt(UserSummaryVolatile),
   });
   const CurrentUserSummary = IDL.Record({
-    'streak' : IDL.Nat16,
     'username' : IDL.Text,
     'is_platform_operator' : IDL.Bool,
     'diamond_membership_status' : DiamondMembershipStatusFull,
-    'next_daily_claim' : TimestampMillis,
     'user_id' : UserId,
     'is_bot' : IDL.Bool,
     'display_name' : IDL.Opt(IDL.Text),
     'avatar_id' : IDL.Opt(IDL.Nat),
     'moderation_flags_enabled' : IDL.Nat32,
-    'chit_balance' : IDL.Int32,
     'is_suspected_bot' : IDL.Bool,
     'suspension_details' : IDL.Opt(SuspensionDetails),
     'is_platform_moderator' : IDL.Bool,
@@ -380,21 +365,6 @@ export const idlFactory = ({ IDL }) => {
       'timestamp' : TimestampMillis,
       'users' : IDL.Vec(UserSummaryV2),
       'current_user' : IDL.Opt(CurrentUserSummary),
-    }),
-  });
-  const UsersV2Args = IDL.Record({
-    'user_groups' : IDL.Vec(
-      IDL.Record({
-        'users' : IDL.Vec(UserId),
-        'updated_since' : TimestampMillis,
-      })
-    ),
-    'users_suspended_since' : IDL.Opt(TimestampMillis),
-  });
-  const UsersV2Response = IDL.Variant({
-    'Success' : IDL.Record({
-      'timestamp' : TimestampMillis,
-      'users' : IDL.Vec(UserSummary),
     }),
   });
   return IDL.Service({
@@ -428,7 +398,6 @@ export const idlFactory = ({ IDL }) => {
         [ChitLeaderboardResponse],
         ['query'],
       ),
-    'claim_daily_chit' : IDL.Func([EmptyArgs], [ClaimDailyChitResponse], []),
     'current_user' : IDL.Func([EmptyArgs], [CurrentUserResponse], ['query']),
     'diamond_membership_fees' : IDL.Func(
         [EmptyArgs],
@@ -531,7 +500,6 @@ export const idlFactory = ({ IDL }) => {
         ['query'],
       ),
     'users' : IDL.Func([UsersArgs], [UsersResponse], ['query']),
-    'users_v2' : IDL.Func([UsersV2Args], [UsersV2Response], ['query']),
   });
 };
 export const init = ({ IDL }) => { return []; };